+++ bee-network/examples/common/mod.rs
@@ -1,8 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#![cfg(feature = "full")]
+#![allow(dead_code)]
 
-pub mod error;
-pub mod list;
-pub mod meta;
+pub mod keys_and_ids;