+++ bee-network/src/service/mod.rs
@@ -3,6 +3,7 @@
 
 #![cfg(feature = "full")]
 
-pub mod error;
-pub mod list;
-pub mod meta;
+pub mod command;
+pub mod controller;
+pub mod event;
+pub mod service;