+++ bee-network/src/swarm/mod.rs
@@ -3,6 +3,6 @@
 
 #![cfg(feature = "full")]
 
-pub mod error;
-pub mod list;
-pub mod meta;
+pub mod behavior;
+pub mod builder;
+pub mod protocols;