+++ bee-network/src/tests/common/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod await_events;
+pub mod keys_and_ids;
+pub mod network_config;
+pub mod shutdown;